% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tissues_for_gene.R
\name{get_tissues_for_gene}
\alias{get_tissues_for_gene}
\title{Title}
\usage{
get_tissues_for_gene(gene_symbol_of_interest)
}
\arguments{
\item{gene_symbol_of_interest}{Character vector of gene symbols.}
}
\description{
Get tissues where a gene has eQTLs.
}
